refactor with new audio utilities package dependency
