+++ com.utilities.encoder.ogg/Packages/com.com.utilities.encoder.ogg/Runtime/OggExtensions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec9d67d25eb20de4fbf225dbd338dc80
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
