+++ com.utilities.encoder.ogg/Packages/com.com.utilities.encoder.ogg/Runtime/Utilities.Encoding.OggVorbis.asmdef
@@ -2,7 +2,8 @@
     "name": "Utilities.Encoding.OggVorbis",
     "rootNamespace": "Utilities.Encoding.OggVorbis",
     "references": [
-        "GUID:a6609af893242c7438d701ddd4cce46a"
+        "GUID:a6609af893242c7438d701ddd4cce46a",
+        "GUID:d25c28436b1dcc9408d86f49a0f5210b"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
