+++ com.utilities.encoder.ogg/ProjectSettings/PackageManagerSettings.asset
@@ -24,19 +24,21 @@ MonoBehaviour:
     m_Url: https://packages.unity.com
     m_Scopes: []
     m_IsDefault: 1
-    m_Capabilities: 0
-  - m_Id: scoped:OpenUPM
+    m_Capabilities: 7
+    m_ConfigSource: 0
+  - m_Id: scoped:project:OpenUPM
     m_Name: OpenUPM
     m_Url: https://package.openupm.com
     m_Scopes:
     - com.utilities
     m_IsDefault: 0
     m_Capabilities: 0
+    m_ConfigSource: 4
   m_UserSelectedRegistryName: OpenUPM
   m_UserAddingNewScopedRegistry: 0
   m_RegistryInfoDraft:
     m_Modified: 0
     m_ErrorMessage: 
-    m_UserModificationsInstanceId: -850
-    m_OriginalInstanceId: -852
+    m_UserModificationsInstanceId: -830
+    m_OriginalInstanceId: -834
   m_LoadAssets: 0
