+++ com.utilities.encoder.ogg/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2021.3.6f1
-m_EditorVersionWithRevision: 2021.3.6f1 (7da38d85baf6)
+m_EditorVersion: 2021.3.18f1
+m_EditorVersionWithRevision: 2021.3.18f1 (3129e69bc0c7)
